Added initial data load into db

+++ src/main/resources/db/migration/V1_1_0__prices_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE inditex_code_challenge.prices (
     id int NOT NULL AUTO_INCREMENT,
     start_date datetime NOT NULL,
     end_date datetime NOT NULL,
-    price int NOT NULL,
+    price decimal NOT NULL,
     price_list int NOT NULL,
     product_id bigint NOT NULL,
     brand_id bigint NOT NULL,
